requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncio==3.4.3
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
idna==3.10
kombu==5.5.4
loguru==0.7.3
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.51
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
